requirements/dev.txt
asgiref==3.7.2
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
Django==5.0
django-filter==23.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
flake8==6.1.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Markdown==3.5.1
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.9.0
urllib3==2.1.0
